// packages/monaco-lsp-client/src/types.ts
export type LSPArray = LSPAny[];

/**
 * The LSP any type.
 * Please note that strictly speaking a property with the value `undefined`
 * can't be converted into JSON preserving the property name. However for
 * convenience it is allowed and assumed that all these properties are
 * optional as well.
 * @since 3.17.0
 */
export type LSPAny = LSPObject | LSPArray | string | number | boolean | null;
// ...
export type DocumentSelector = DocumentFilter[];
// ...
export type LSPObject = { [key: string]: LSPAny };
// ...
export interface WorkDoneProgressOptions {
  workDoneProgress?: boolean;
}

/**
 * General text document registration options.
 */
export interface TextDocumentRegistrationOptions {
  /**
   * A document selector to identify the scope of the registration. If set to null
   * the document selector provided on the client side will be used.
   */
  documentSelector: DocumentSelector | null;
}
// ...
export interface Command {
  /**
   * Title of the command, like `save`.
   */
  title: string;
  /**
   * The identifier of the actual command handler.
   */
  command: string;
  /**
   * Arguments that the command handler should be
   * invoked with.
   */
  arguments?: LSPAny[];
}
// ...
export interface CodeLens {
  /**
   * The range in which this code lens is valid. Should only span a single line.
   */
  range: Range;
  /**
   * The command this code lens represents.
   */
  command?: Command;
  /**
   * A data entry field that is preserved on a code lens item between
   * a {@link CodeLensRequest} and a {@link CodeLensResolveRequest}
   */
  data?: LSPAny;
}

/**
 * Registration options for a {@link CodeLensRequest}.
 */
export interface CodeLensRegistrationOptions
  extends TextDocumentRegistrationOptions, CodeLensOptions {}
// ...
export interface Range {
  /**
   * The range's start position.
   */
  start: Position;
  /**
   * The range's end position.
   */
  end: Position;
}
// ...
export interface Position {
  /**
   * Line position in a document (zero-based).
   *
   * If a line number is greater than the number of lines in a document, it defaults back to the number of lines in the document.
   * If a line number is negative, it defaults to 0.
   */
  line: number;
  /**
   * Character offset on a line in a document (zero-based).
   *
   * The meaning of this offset is determined by the negotiated
   * `PositionEncodingKind`.
   *
   * If the character value is greater than the line length it defaults back to the
   * line length.
   */
  character: number;
}
// ...
export interface CodeLensOptions extends WorkDoneProgressOptions {
  /**
   * Code lens has a resolve provider as well.
   */
  resolveProvider?: boolean;
}